+++ doc/source/doxygen-docs/changelog.md
@@ -23,6 +23,8 @@
     - mrpt::containers::yaml and libfyaml updated to latest version (more memory efficient parser).
   - \ref mrpt_core_grp
     - New base virtual interface class mrpt::Stringifyable unifying the asString() method already offered by many MRPT classes.
+  - \ref mrpt_img_grp
+    - **[API change]** mrpt::img::TCamera methods changed to allow defining fish-eye camera models too.
   - \ref mrpt_io_grp
     - GZIP compressed streams now also support open and append. See new mrpt::io::CFileGZOutputStream::open() signature.
     - New enum mrpt::io::OpenMode for clearer-to-read code.